interfaces refactor
