+++ src/components/Insertion/types.ts
@@ -1,14 +1,3 @@
-export interface UserLogIn {
-  email: string;
-  password: string;
-}
-
-export interface UserSignIn {
-  email: string;
-  password: string;
-  confirmPassword: string;
-}
-
 export interface InsertionData {
   instrumentType: string;
   model: string;
