+++ src/components/UserLogin/types.ts
@@ -1,0 +1,4 @@
+export interface UserLogIn {
+  email: string;
+  password: string;
+}
