+++ src/components/UserSignIn/types.ts
@@ -1,0 +1,5 @@
+export interface UserSignIn {
+  email: string;
+  password: string;
+  confirmPassword: string;
+}
